Struct string_lines::StringLines

pub struct StringLines {
    // some fields omitted
}

Raw persistent database for storing string lines.

Methods

impl StringLines

fn open<P: AsRef<Path>>(path: P) -> Result<StringLines>

Attempts to open a file in read-write mode.

fn push(&mut self, s: &str) -> Result<()>

Appends an element to the back of a collection.

fn pop(&mut self) -> Result<Option<String>>

Removes the last element from a collection and returns it, or None if it is empty.

fn clear(&mut self) -> Result<()>

Returns the number of elements in the collection.